Add 'back to all auctions' link
* Doing this instead of breadcrumbs
* Closes https://github.com/18F/micropurchase/issues/694
* Also update routes to make them more-railsy
