+++ app/assets/stylesheets/components/_bids.scss
@@ -19,9 +19,4 @@
   .cancel {
     margin-right: 2rem;
   }
-
-  .button-wrapper {
-    margin: 0 auto;
-    width: 600px;
-  }
 }
