--- app/views/auctions/_winning_bid_amount.html.erb
@@ -1,4 +1,0 @@
-<div class="issue-bids-info-item">
-  <span>Winning bid: </span>
-  <%= auction.highlighted_bid_amount_as_currency %>
-</div>
