+++ app/views/bid_confirmations/create.html.erb
@@ -1,0 +1,23 @@
+<h1>Confirm Your Bid</h1>
+
+<h1 class='issue-title'><%= @confirm_bid.title %></h1>
+<p class='issue-description'>
+  <%= @confirm_bid.html_description %>
+</p>
+
+<h3><%= @confirm_bid.time_left %></h3>
+
+<%= form_for @confirm_bid.bid,
+  url: auction_bids_path(@confirm_bid.auction_id),
+  html: { class: 'bid-form' } do |f| %>
+  <%= f.hidden_field :amount %>
+
+  <div class="usa-width-one-whole">
+    <h1>Your bid: $<%= @confirm_bid.bid_amount %></h1>
+
+    <%= link_to 'Cancel',
+      auction_path(@confirm_bid.auction_id),
+      class: 'usa-button usa-button-gray cancel' %>
+    <%= f.submit 'Confirm', class: 'usa-button submit' %>
+  </div>
+<% end %>
