--- app/views/bids/_auction_details.html.erb
@@ -1,10 +1,0 @@
-<div class='usa-width-one-whole issue-list-item'>
-  <h1 class='issue-title'><%= bid_view_model.title %></h1>
-  <p class='issue-description'>
-    <%= bid_view_model.html_description %>
-  </p>
-
-  <%= render partial: bid_view_model.time_left_partial,
-    locals: { bid_view_model: bid_view_model } %>
-</div>
-<hr>
