--- app/views/bids/_time_left.html.erb
@@ -1,1 +1,0 @@
-<h3>Ends in <%= bid_view_model.distance_of_time %></h3>
