--- app/views/bids/confirm.html.erb
@@ -1,20 +1,0 @@
-<h1>Confirm Your Bid</h1>
-
-<%= render partial: 'auction_details', locals: { bid_view_model: @confirm_bid } %>
-
-<%= form_for @confirm_bid.bid,
-  url: auction_bids_path(@confirm_bid.auction_id),
-  html: { class: 'bid-form' } do |f| %>
-  <%= f.hidden_field :amount %>
-
-  <div class="usa-width-one-whole">
-    <h1 class='center'>Confirm your bid: $<%= @confirm_bid.bid_amount %></h1>
-  </div>
-
-  <div class='button-wrapper'>
-    <%= link_to 'Cancel',
-      root_path,
-      class: 'usa-button usa-button-big usa-button-gray cancel' %>
-    <%= f.submit 'Confirm', class: 'usa-button-big usa-button submit' %>
-  </div>
-<% end %>
