+++ app/views/bids/index.html.erb
@@ -1,12 +1,4 @@
-<% content_for :title do %>18F Micro-purchase - Your bids<% end %>
-<% content_for :description do %>View your bids across Micro-purchase auctions.<% end %>
-
 <h1>My Bids</h1>
-
-<%= link_to '<icon class="fa fa-angle-double-left"></icon> Back to open issues'.html_safe,
-  root_path,
-  class: "link-highlighted" %>
-
 
 <div class="usa-grid-full">
   <table id="table-admins-auction" class="admin-auctions-table">
